% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_med_load.R
\name{calculate_med_load}
\alias{calculate_med_load}
\title{Calculate med load}
\usage{
calculate_med_load()
}
\description{
This function allow you to start the process of med load/response calculations
}
\examples{
calculate_seizure_load()
}
\keyword{load}
\keyword{med}

