% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_med_response.R
\name{calculate_med_response}
\alias{calculate_med_response}
\title{Calculate med response}
\usage{
calculate_med_response(patient, data, y, calculated, first, last)
}
\arguments{
\item{patient}{A patient identifier}

\item{data}{The med raw data table}

\item{y}{The med load table}

\item{calculated}{The med load table including baseline days (to be uploaded to the MySQL database)}

\item{first}{Patient's first name}

\item{last}{Patient's last name}
}
\description{
This function allow you to calculate med responses (i.e. med response, % med free, med score, etc.).
You can upload this data to a MySQL database if you choose.
}
\keyword{med}
\keyword{response}

