% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_outcome.R
\name{calculate_outcome}
\alias{calculate_outcome}
\title{Calculate outcome}
\usage{
calculate_outcome()
}
\description{
This function allow you to calculate a patient's outcome (i.e. it creates the OUTCOME_DATA_CLINICAL
file and the MED_SEIZURE_OUTCOME_BAR_TABLE files, as well as creates bar graphs for seizure score,
med score, and outcome for each period of 30 days)
}
\examples{
calculate_outcome()
}
\keyword{outcome}

