% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ranks.R
\name{calculate_ranks}
\alias{calculate_ranks}
\title{Calculate Ranks}
\usage{
calculate_ranks(n, m)
}
\arguments{
\item{n}{Ranking subtable}

\item{m}{Maximum rank value}
}
\value{
Ordered ranking subtable
}
\description{
This function will take a portion of a ranking table and order it in
increasing numerical order based on ranking value
}
\keyword{calculate}
\keyword{rank}

