% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_seizure_load.R
\name{calculate_seizure_load}
\alias{calculate_seizure_load}
\title{Calculate seizure load}
\usage{
calculate_seizure_load()
}
\description{
This function allow you to start the process of seizure load/response calculations
}
\examples{
calculate_seizure_load()
}
\keyword{load}
\keyword{seizure}

