% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_seizure_response.R
\name{calculate_seizure_response}
\alias{calculate_seizure_response}
\title{Calculate seizure response}
\usage{
calculate_seizure_response(patient, data, x, sourcedata, ranking)
}
\arguments{
\item{patient}{A patient identifier}

\item{data}{The seizure raw data table}

\item{x}{The seizure load table}

\item{sourcedata}{The seizure load table including baseline days (to be uploaded to the MySQL database)}

\item{ranking}{The seizure ranking table}
}
\description{
This function allow you to calculate seizure responses (i.e. seizure response, % seizure free, seizure score, etc.).
You can upload this data to a MySQL database if you choose.
}
\keyword{response}
\keyword{seizure}

