% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLINIC_VISIT_SOURCE.R
\docType{data}
\name{CLINIC_VISIT_SOURCE}
\alias{CLINIC_VISIT_SOURCE}
\title{Clinic Visit Source}
\format{Source data for patients' clinic visit information}
\usage{
CLINIC_VISIT_SOURCE
}
\description{
Clinic Visit Source
}
\keyword{clinic}
\keyword{source}
\keyword{visit}

