% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEMOGRAPHICS_SOURCE.R
\docType{data}
\name{DEMOGRAPHICS_SOURCE}
\alias{DEMOGRAPHICS_SOURCE}
\title{Demographics Source}
\format{Source data for patients' demographics information}
\usage{
DEMOGRAPHICS_SOURCE
}
\description{
Demographics Source
}
\keyword{demographics}
\keyword{source}

