% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makelowercase.R
\name{makelowercase}
\alias{makelowercase}
\title{Make Lowercase}
\usage{
makelowercase(x)
}
\arguments{
\item{x}{Column of the raw data table (either severity, length, type, variable, or cluster)}
}
\value{
Column of the raw data table with all alphabetic data in lowercase (if it wasn't already)
}
\description{
This function ensure that all alphabetic data in the raw seizure data table is lowercase
}
\keyword{lowercase}
\keyword{make}

