% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_calculate.R
\name{med_calculate}
\alias{med_calculate}
\title{Med calculate}
\usage{
med_calculate(x, n, l, baseline, therapy, patient, mrnumber, type, number)
}
\arguments{
\item{x}{Med load table}

\item{n}{Number that signifies which column number corresponse with the MED_NUMBER_DAY column}

\item{l}{Number that signifies which column number corresponds with the MED_LOAD_DAY column}

\item{baseline}{Subset of load table corresponding with baseline days}

\item{therapy}{Subset of load table corresponding with therapy days}

\item{patient}{Four-letter patient initials}

\item{mrnumber}{Medical record number object}

\item{type}{DAY_TYPE column object}

\item{number}{MED_NUMBER_DAY/MED_LOAD_DAY data frame object}
}
\value{
med response/score calculations
}
\description{
This function will perform the calculations of the values for percent med free (on baseline days
and therapy days), med response (daily and for every 30 days), percent med
free response, med score, med number response (daily and for every 30 days),
and med number score
}
\keyword{med}
\keyword{response}

