% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MED_RANKING_SOURCE.R
\docType{data}
\name{MED_RANKING_SOURCE}
\alias{MED_RANKING_SOURCE}
\title{Med Ranking Source}
\format{Source data for patients' med ranking table}
\usage{
MED_RANKING_SOURCE
}
\description{
Med Ranking Source
}
\keyword{med}
\keyword{ranking}
\keyword{source}

