% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_sums.R
\name{missing_sums}
\alias{missing_sums}
\title{Missing sums}
\usage{
missing_sums(x, y)
}
\arguments{
\item{x}{Subset of the seizure load table}

\item{y}{Subset of the DATA_QUALITY_S column of the raw data table that matches with the rows of the seizure load table being used by this function}
}
\value{
The seizure load table with any missing values replaced with the appropriate seizure number/load value according to the missing data rules
}
\description{
This function is used to assign ranking values to days with no observed data in the raw
data table, i.e. missing days. The rules are as follows: for each block of missing data,
count the number of consecutive missing days. Split this block up into two portions, either
of equal length if this block is of even length, or the upper subblock one longer than the
lower subblock if the block is of odd length. To fill the upper subblock, we will draw the
load and number values of the amount of days with observed data numbering the length of the
upper subblock immediately preceding the upper subblock and use those as the load and number
values for the subblock. Similar process is done for the lower subblock, looking at the days
with observed data immediately following the subblock. If the beginning or the end of the rows
of data are reached before the required number of rows of observed data have been found, repeat
the data found as many times as necessary to fill the subblock of missing data
}
\keyword{range}
\keyword{rank}

