% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATIENT1_ANTHROPOMETRICS_SOURCE.R
\docType{data}
\name{PATIENT1_ANTHROPOMETRICS_SOURCE}
\alias{PATIENT1_ANTHROPOMETRICS_SOURCE}
\title{PATIENT1 Anthropometrics Source}
\format{Source data for PATIENT1's anthropometrics data}
\usage{
PATIENT1_ANTHROPOMETRICS_SOURCE
}
\description{
PATIENT1 Anthropometrics Source
}
\keyword{anthropometrics}
\keyword{source}

