% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATIENT1_MED_DATA_SOURCE.R
\docType{data}
\name{PATIENT1_MED_DATA_SOURCE}
\alias{PATIENT1_MED_DATA_SOURCE}
\title{PATIENT1 Med Data Source}
\format{Source data for PATIENT1's med data}
\usage{
PATIENT1_MED_DATA_SOURCE
}
\description{
PATIENT1 Med Data Source
}
\keyword{data}
\keyword{med}
\keyword{source}

