% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATIENT1_SEIZURE_DATA_SOURCE.R
\docType{data}
\name{PATIENT1_SEIZURE_DATA_SOURCE}
\alias{PATIENT1_SEIZURE_DATA_SOURCE}
\title{PATIENT1 Seizure Data Source}
\format{Source data for PATIENT1's seizure data}
\usage{
PATIENT1_SEIZURE_DATA_SOURCE
}
\description{
PATIENT1 Seizure Data Source
}
\keyword{data}
\keyword{seizure}
\keyword{source}

