% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATIENT1_SEIZURE_RANKING_SOURCE.R
\docType{data}
\name{PATIENT1_SEIZURE_RANKING_SOURCE}
\alias{PATIENT1_SEIZURE_RANKING_SOURCE}
\title{PATIENT1 Seizure Ranking Source}
\format{Source data for PATIENT1's seizure ranking}
\usage{
PATIENT1_SEIZURE_RANKING_SOURCE
}
\description{
PATIENT1 Seizure Ranking Source
}
\keyword{ranking}
\keyword{seizure}
\keyword{source}

