% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankwithranges.R
\name{rankwithranges}
\alias{rankwithranges}
\title{Rank With Ranges}
\usage{
rankwithranges(x, b)
}
\arguments{
\item{x}{Observed value in the raw data table}

\item{b}{First two columns of the ranking subtable}
}
\value{
ranking value to be assigned to this observed data value
}
\description{
This function draws the appropriate ranking value from the appropriate ranking subtable
based on the value found in the raw data table (for variables that have observations that
are observed in ranges, i.e. length and cluster)
}
\keyword{range}
\keyword{rank}

