% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seizure_calculate.R
\name{seizure_calculate}
\alias{seizure_calculate}
\title{Seizure calculate}
\usage{
seizure_calculate(x, n, l, baseline, therapy, patient, mrnumber, type, quality,
  number)
}
\arguments{
\item{x}{Seizure load table}

\item{n}{Number that signifies which column number corresponse with the SEIZURE_NUMBER_DAY column}

\item{l}{Number that signifies which column number corresponds with the SEIZURE_LOAD_DAY column}

\item{baseline}{Subset of load table corresponding with baseline days}

\item{therapy}{Subset of load table corresponding with therapy days}

\item{patient}{Four-letter patient initials}

\item{mrnumber}{Medical record number object}

\item{type}{DAY_TYPE column object}

\item{quality}{DATA_QUALITY_S column object}

\item{number}{SEIZURE_NUMBER_DAY/SEIZURE_LOAD_DAY data frame object}
}
\value{
Seizure response/score calculations
}
\description{
This function will perform the calculations of the values for percent seizure free (on baseline days
and therapy days), seizure response (daily and for every 30 days), percent seizure
free response, seizure score, seizure number response (daily and for every 30 days),
and seizure number score
}
\keyword{response}
\keyword{seizure}

