% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seizure_sum.R
\name{seizure_sum}
\alias{seizure_sum}
\title{Seizure sum}
\usage{
seizure_sum(x, a, b, c, d, e)
}
\arguments{
\item{x}{Raw data table}

\item{a}{Severity ranking subtable}

\item{b}{Length ranking subtable}

\item{c}{Type ranking subtable}

\item{d}{Variable ranking subtable}

\item{e}{Cluster ranking subtable}
}
\value{
A table that stores the ranking values assigned to each observed data value (NA for missing data)
}
\description{
This function takes the raw seizure table and assign to each observed value within
within that table a ranking value based on the ranking assignments given within
the ranking table. If a row in the data table is a missing day (i.e. DATA_QUALITY_S = 3),
then assign NA values to that row for now, to be dealt with later.
}
\keyword{range}
\keyword{rank}

