% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_interp.R
\name{weight_interp}
\alias{weight_interp}
\title{Seizure calculate}
\usage{
weight_interp(Anthropometrics)
}
\arguments{
\item{Anthropometrics}{The anthropometrics source table, to be turned into a table with interpolated weights}
}
\value{
Anthropometrics table with interpolated weights
}
\description{
This function will take the weights observed at the clinic dates and interpolate the weights of the patient 
on non-clinic days, based on linear interpolation
}
\keyword{interpolation}
\keyword{weight}

